Moving data microservice to cmd
