+++ cmd/core-data/res/configuration-docker.json
