+++ cmd/core-data/res/configuration.json
